mds: support for SAF SaTimeT timeout variable in MDS transport [#1658] Issue : In current implementation their is a mismatch of variable sizes between SAF API accepting SaTimeT variable type for timeout ( 64 bit) and MDS transport implementation accepting (32 bit) timeout variable , the MDS transport should adopt the size of the timeout parameter from 32 bits to 64 bits.
Fix :
Now MDS provide interface to accept SaTimeT type timeout variable from SAF APIS

+++ osaf/libs/core/mds/include/mds_dl_api.h
@@ -32,9 +32,9 @@
 
 uint32_t mds_lib_req(NCS_LIB_REQ_INFO *req);
 int mds_auth_server_connect(const char *name, MDS_DEST mds_dest,
-                            int svc_id, int timeout);
+                            int svc_id, int64_t timeout);
 int mds_auth_server_disconnect(const char *name, MDS_DEST mds_dest,
-                               int svc_id, int timeout);
+                               int svc_id, int64_t timeout);
 int mds_auth_server_create(const char *name);
 
 #endif
